# Inline callsite of invoke when possible (`invoke` improvement No. 1)