--- a/src/gf.c
+++ b/src/gf.c
@@ -1675,6 +1675,89 @@ DLLEXPORT jl_value_t *jl_gf_invoke_lookup(jl_function_t *gf, jl_datatype_t *type
     return (jl_value_t*)m;
 }
 
+static jl_function_t*
+invoke_specialize(jl_methlist_t *m, jl_methtable_t *mt, jl_tupletype_t *tt)
+{
+    jl_svec_t *tpenv = jl_emptysvec;
+    jl_tupletype_t *newsig = NULL;
+    JL_GC_PUSH2(&tpenv, &newsig);
+
+    if (m->invokes == (void*)jl_nothing) {
+        m->invokes = new_method_table(mt->name);
+        gc_wb(m, m->invokes);
+        update_max_args(m->invokes, tt);
+        // this private method table has just this one definition
+        jl_method_list_insert(&m->invokes->defs, m->sig, m->func,
+                              m->tvars, 0, 0, (jl_value_t*)m->invokes);
+    }
+
+    newsig = m->sig;
+
+    if (m->tvars != jl_emptysvec) {
+        jl_value_t *ti = lookup_match((jl_value_t*)tt, (jl_value_t*)m->sig,
+                                      &tpenv, m->tvars);
+        assert(ti != (jl_value_t*)jl_bottom_type);
+        (void)ti;
+        // don't bother computing this if no arguments are tuples
+        for (size_t i = 0;i < jl_nparams(tt);i++) {
+            if (jl_is_tuple_type(jl_tparam(tt, i))) {
+                newsig = (jl_tupletype_t*)jl_instantiate_type_with(
+                    (jl_value_t*)m->sig,
+                    jl_svec_data(tpenv),
+                    jl_svec_len(tpenv) / 2);
+                break;
+            }
+        }
+    }
+    jl_function_t *mfunc =
+        cache_method(m->invokes, tt, m->func, newsig, tpenv, m->isstaged);
+    JL_GC_POP();
+    return mfunc;
+}
+
+// compile-time method lookup
+jl_function_t*
+jl_gf_invoke_get_specialization(jl_function_t *gf, jl_tupletype_t *types,
+                                jl_tupletype_t *tt)
+{
+    assert(jl_is_gf(gf));
+    jl_methtable_t *mt = jl_gf_mtable(gf);
+    jl_methlist_t *m = (jl_methlist_t*)jl_gf_invoke_lookup(gf, types);
+    size_t i;
+
+    if ((jl_value_t*)m == jl_nothing) {
+        return NULL;
+    }
+
+    // now we have found the matching definition.
+    // next look for or create a specialization of this definition.
+
+    jl_function_t *mfunc;
+    if (m->invokes == (void*)jl_nothing)
+        mfunc = jl_bottom_func;
+    else
+        mfunc = jl_method_table_assoc_exact_by_type(m->invokes, tt);
+    JL_GC_PUSH1(&mfunc);
+    if (mfunc != jl_bottom_func) {
+        if (mfunc->linfo == NULL || mfunc->linfo->inInference ||
+            mfunc->linfo->inCompile) {
+            JL_GC_POP();
+            return NULL;
+        }
+    } else {
+        mfunc = invoke_specialize(m, mt, tt);
+    }
+    if (mfunc->linfo->functionObject == NULL) {
+        if (mfunc->fptr != &jl_trampoline) {
+            JL_GC_POP();
+            return NULL;
+        }
+        jl_compile(mfunc);
+    }
+    JL_GC_POP();
+    return mfunc;
+}
+
 // invoke()
 // this does method dispatch with a set of types to match other than the
 // types of the actual arguments. this means it sometimes does NOT call the
@@ -1721,38 +1804,9 @@ jl_value_t *jl_gf_invoke(jl_function_t *gf, jl_tupletype_t *types,
         }
     }
     else {
-        jl_svec_t *tpenv=jl_emptysvec;
-        jl_tupletype_t *newsig=NULL;
-        jl_tupletype_t *tt=NULL;
-        JL_GC_PUSH3(&tpenv, &newsig, &tt);
-        tt = arg_type_tuple(args, nargs);
-        if (m->invokes == (void*)jl_nothing) {
-            m->invokes = new_method_table(mt->name);
-            gc_wb(m, m->invokes);
-            update_max_args(m->invokes, tt);
-            // this private method table has just this one definition
-            jl_method_list_insert(&m->invokes->defs,m->sig,m->func,m->tvars,0,0,(jl_value_t*)m->invokes);
-        }
-
-        newsig = m->sig;
-
-        if (m->tvars != jl_emptysvec) {
-            jl_value_t *ti =
-                lookup_match((jl_value_t*)tt, (jl_value_t*)m->sig, &tpenv, m->tvars);
-            assert(ti != (jl_value_t*)jl_bottom_type);
-            (void)ti;
-            // don't bother computing this if no arguments are tuples
-            for(i=0; i < jl_nparams(tt); i++) {
-                if (jl_is_tuple_type(jl_tparam(tt,i)))
-                    break;
-            }
-            if (i < jl_nparams(tt)) {
-                newsig = (jl_tupletype_t*)jl_instantiate_type_with((jl_value_t*)m->sig,
-                                                                   jl_svec_data(tpenv),
-                                                                   jl_svec_len(tpenv)/2);
-            }
-        }
-        mfunc = cache_method(m->invokes, tt, m->func, newsig, tpenv, m->isstaged);
+        jl_tupletype_t *tt = arg_type_tuple(args, nargs);
+        JL_GC_PUSH1(&tt);
+        mfunc = invoke_specialize(m, mt, tt);
         JL_GC_POP();
     }