@yuyichao 2015-04-23
A slightly more complete test (doesn't include the `@code_llvm` output of the master version since it's way too long....)

```julia
@noinline f1(a, b) = a + b

g1_1() = f1(1, 2)
g1_2() = invoke(f1, Tuple{Any, Any}, 1, 2)

@inline t1() = Tuple{Any, Any}
@noinline t2() = Tuple{Any, Any}

g1_3() = invoke(f1, t1(), 1, 2)
g1_4() = invoke(f1, t2(), 1, 2)

@assert g1_1() == g1_2() == g1_3() == g1_4()

@code_llvm g1_1()
@code_llvm g1_2()
@code_llvm g1_3()
@code_llvm g1_4()

function timing(f, args...)
    println(f, args)
    f(args...)
    gc()
    @time for i in 1:10000000
        f(args...)
    end
    gc()
end

timing(g1_1)
timing(g1_2)
timing(g1_3)
timing(g1_4)
```

master without `@code_llvm` output
```llvm
g1_1()
elapsed time: 0.308610973 seconds (0 bytes allocated)
g1_2()
elapsed time: 1.89693006 seconds (1449 MB allocated, 1.36% gc time in 66 pauses with 0 full sweep)
g1_3()
elapsed time: 1.892066647 seconds (1449 MB allocated, 1.41% gc time in 66 pauses with 0 full sweep)
g1_4()
elapsed time: 1.883954768 seconds (1449 MB allocated, 1.41% gc time in 66 pauses with 0 full sweep)
```

with this PR with `@code_llvm` output
```llvm
define i64 @julia_g1_1_43914() {
top:
  %0 = call i64 @julia_f1_43903(i64 1, i64 2)
  ret i64 %0
}

define i64 @julia_g1_2_43917() {
top:
  %0 = call i64 @julia_f1_43903(i64 1, i64 2)
  ret i64 %0
}

define i64 @julia_g1_3_43918() {
top:
  %0 = call i64 @julia_f1_43903(i64 1, i64 2)
  ret i64 %0
}

define i64 @julia_g1_4_43919() {
top:
  %0 = call %jl_value_t* @julia_t2_43907()
  %1 = call i64 @julia_f1_43903(i64 1, i64 2)
  ret i64 %1
}
g1_1()
elapsed time: 0.319919904 seconds (0 bytes allocated)
g1_2()
elapsed time: 0.372261468 seconds (0 bytes allocated)
g1_3()
elapsed time: 0.323118056 seconds (0 bytes allocated)
g1_4()
elapsed time: 1.10516864 seconds (1449 MB allocated, 2.79% gc time in 66 pauses with 0 full sweep)
```