@pao 2015-04-24
The 32-bit error on Travis is indeed tupocalypse fallout and it's being worked on.