@yuyichao 2015-04-24
Was the issue before repeatable?
I asked whether it's a connection issue because both build [1](https://ci.appveyor.com/project/StefanKarpinski/julia/build/1.0.4069/job/wwdnwkpxha87m2me) [2](https://ci.appveyor.com/project/StefanKarpinski/julia/build/1.0.4066/job/o3777iciq624y3jd) stops responding at ~11mins at different places.....