@yuyichao 2015-04-25
@vtjnash So I fixed two of the three issues you brought up.

For the thrid one, it seems that `emit_call` made a lot of assumptions of what `emit_known_call` might do and I feel like I would have too pass back too much information (`nargs`/`args`, `f`) without gaining much.

I've tried to replace the `emit_nthptr_recast` with just `mfunc->linfo->functionObject` and got the same `SegFault` with before.