@yuyichao 2015-04-25
And somehow the `x86_64` test freeze again............. :(