@yuyichao 2015-05-30
I think I've finally understand how these code works (thanks to https://github.com/JuliaLang/julia/pull/11439) and why it was crashing before.

As @carnaval guessed, it was related to GC and as I noted in the commit https://github.com/JuliaLang/julia/commit/8dce0332361d22cfbe9c39ecc5b04c78997484fb#diff-6d4d21428a67320600faf5a1a9f3a16aR2437, the function object can be collected when the function object is passed to the jlcall function. This shouldn't be an issue now and not even with https://github.com/JuliaLang/julia/pull/11439 since a non-wrapper jlcall specialization never accesses the function object (first) argument. I'm still not entirely happy that there's a useless and possibly invalid pointer in the generated code though...................