@yuyichao 2016-09-11
Replaced by https://github.com/JuliaLang/julia/pull/18444